Fixed map control and map appearance
This is both on mobile and desktop.  We rely on Google Maps JavaScript
API to select the best experience.  We explicitly disable only street
view control and map type control which doen't make sense in our
experience.

Scaling on mobile was changed to 1.0 which gives better result than
device width.